+++ js/tabview.js
@@ -46,7 +46,8 @@
             return (['audio/flac', 'audio/mp4', 'audio/mpeg', 'audio/ogg', 'audio/wav',
                 'image/gif', 'image/heic', 'image/jpeg', 'image/png', 'image/tiff', 'image/x-dcraw',
                 'video/3gpp', 'video/dvd', 'video/mp4', 'video/mpeg', 'video/quicktime',
-                'video/webm', 'video/x-flv', 'video/x-matroska', 'video/x-msvideo'].indexOf(mimetype) > -1);
+                'video/webm', 'video/x-flv', 'video/x-matroska', 'video/x-msvideo',
+                'application/zip'].indexOf(mimetype) > -1);
         },
 
         formatValue: function(value) {